refactor: address review — org-mint endpoint, comment cleanup, extractions
- Move the user-context token mint off the project controller to
  `POST /v2/organizations/{organizationId}/apps/{installId}/token` (org-wide
  token, no projectId); AppTokenModel moves out of the project package.
- Extract app-token authentication out of AuthenticationFilter into
  AppTokenAuthenticator, and the app project-binding out of
  ProjectContextService into AppProjectContextBinder, keeping both host
  classes minimal.
- Make AppSecretService.expireOthers private; drop the always-false `native`
  field from the app-self install model.
- Remove narrating inline comments across the app changeset, keeping only the
  load-bearing security/ordering notes.
