+++ backend/data/src/main/kotlin/io/tolgee/dtos/apps/AppEnabledProjectDto.kt
@@ -1,6 +1,5 @@
 package io.tolgee.dtos.apps
 
-/** A project an app install is enabled for, with the organization owning that project. */
 data class AppEnabledProjectDto(
   val id: Long,
   val name: String,
