+++ backend/data/src/main/kotlin/io/tolgee/service/apps/AppInstallService.kt
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional
 /**
  * Orchestrates without a transaction: the manifest fetch reaches an app-controlled host that may
  * stall for seconds, and holding a pooled DB connection across it would let a handful of slow hosts
- * exhaust the pool. Writes are delegated to [AppInstallPersister].
+ * exhaust the pool.
  */
 @Service
 class AppInstallService(
@@ -29,26 +29,16 @@ class AppInstallService(
   data class RegisterResult(
     val install: AppInstall,
     val app: AppService.AppSummary,
-    /** Always null here: installing an already-registered app discloses no app-level credentials. */
     val appCredentials: AppService.AppCredentials?,
   )
 
   data class RegisterAppResult(
     val app: AppService.AppSummary,
     val appEntityId: Long,
-    /** The app-level credentials, disclosed only in this response to registering the app. */
     val appCredentials: AppService.AppCredentials?,
-    /** Non-null when the app was also installed for the owner (the default). */
     val install: AppInstall?,
   )
 
-  /**
-   * Registers the app for the organization - making it the app's owner - and, unless [install] is
-   * false, installs it, all in one transaction. Registering an app that is already registered is
-   * refused with [Message.APP_ALREADY_REGISTERED]; installing an existing app is the separate
-   * install endpoint. This runs outside a transaction so the concurrent-duplicate race can be caught
-   * after the write's own transaction has fully rolled back.
-   */
   fun register(
     organization: Organization,
     manifestUrl: String,
@@ -64,18 +54,6 @@ class AppInstallService(
     }
   }
 
-  /**
-   * Installs an app that is already registered on this server, refusing with
-   * [io.tolgee.exceptions.AppNotRegisteredException] when it is not - installing must never register
-   * an app behind the caller's back, because registering hands out app-level credentials and makes
-   * the caller's organization the app's owner.
-   *
-   * The install snapshot is taken from the **registered** manifest URL, not from the one the caller
-   * pasted: an app is identified by the id in its manifest, so a lookalike manifest served elsewhere
-   * would otherwise decide the scopes and base URL of an install of somebody else's app. The consent
-   * hash is likewise checked against that authoritative manifest - the one actually installed - not
-   * against the pasted fetch.
-   */
   fun install(
     organization: Organization,
     manifestUrl: String,
@@ -109,16 +87,10 @@ class AppInstallService(
     return appManifestFetcher.fetch(manifestUrl)
   }
 
-  /** The SHA-256 hex of the manifest as fetched, so the consent preview and the write agree on it. */
   fun manifestHash(fetched: AppManifestFetcher.FetchResult): String {
     return DigestUtils.sha256Hex(fetched.rawJson)
   }
 
-  /**
-   * Rejects a manifest whose bytes changed between the consent preview and this write, so an app
-   * cannot widen the scopes it requests after they were approved. Skipped when the caller sends no
-   * hash (nothing was previewed to compare against).
-   */
   private fun verifyManifestUnchanged(
     expectedHash: String?,
     fetched: AppManifestFetcher.FetchResult,
@@ -129,10 +101,6 @@ class AppInstallService(
     }
   }
 
-  /**
-   * Uninstalls the app from one organization. The app itself stays registered and every other
-   * organization's install is untouched.
-   */
   fun remove(
     organizationId: Long,
     installId: Long,
@@ -145,7 +113,6 @@ class AppInstallService(
     return appInstallRepository.findAllByOrganizationId(organizationId)
   }
 
-  /** The organizations that currently have the app installed, for the admin installations view. */
   @Transactional(readOnly = true)
   fun findInstallingOrganizations(
     appEntityId: Long,
@@ -163,17 +130,11 @@ class AppInstallService(
     return appInstallRepository.findByOrganizationIdAndId(organizationId, installId)
   }
 
-  /** Install by id alone, for the auth filter; tenant safety comes from the later enablement check. */
   @Transactional(readOnly = true)
   fun findForAppAuth(installId: Long): AppInstall? {
     return appInstallRepository.findWithAppById(installId)
   }
 
-  /**
-   * Resolves an install together with the identity an install-context request runs as — the
-   * install's own [AppInstall.principal] (see its definition for why, and for where the install's
-   * capability comes from instead).
-   */
   @Transactional(readOnly = true)
   fun resolveForAppAuth(installId: Long): AppAuthResolution? {
     val install = appInstallRepository.findWithAppById(installId) ?: return null
@@ -182,22 +143,14 @@ class AppInstallService(
 
   data class AppAuthResolution(
     val install: AppInstall,
-    /** The install acting as itself. */
     val principal: UserAccountDto,
   )
 
-  /** Every installation of the app, across organizations, for the app's own discovery call. */
   @Transactional(readOnly = true)
   fun findAllByRegisteredApp(appEntityId: Long): List<AppInstall> {
     return appInstallRepository.findAllByRegisteredAppId(appEntityId)
   }
 
-  /**
-   * The app's own install [installId], for an app minting a token with its app-level credentials.
-   *
-   * Returns null both when no such install exists and when it belongs to a different app, so an
-   * authenticated app cannot use this to learn which install ids exist outside its own.
-   */
   @Transactional(readOnly = true)
   fun findOwnInstall(
     appEntityId: Long,
@@ -208,7 +161,6 @@ class AppInstallService(
     return install
   }
 
-  /** The app itself, for authenticating an app-level token in the auth filter. */
   @Transactional(readOnly = true)
   fun findAppForAppAuth(appEntityId: Long): App? {
     return appService.findForAppAuth(appEntityId)
