+++ backend/data/src/main/kotlin/io/tolgee/service/apps/AppService.kt
@@ -30,10 +30,6 @@ class AppService(
     return appRepository.findByAppId(appId) ?: throw AppNotRegisteredException(appId)
   }
 
-  /**
-   * The app, provided [organizationId] owns it. An organization that merely installed somebody
-   * else's app must not reach it here - administering an app is the owner's alone.
-   */
   @Transactional(readOnly = true)
   fun getOwned(
     organizationId: Long,
@@ -48,7 +44,6 @@ class AppService(
     return appRepository.findById(appEntityId).orElseThrow { NotFoundException(Message.APP_NOT_FOUND) }
   }
 
-  /** App by id alone, for authenticating an app-level token in the auth filter. */
   @Transactional(readOnly = true)
   fun findForAppAuth(appEntityId: Long): App? {
     return appRepository.findById(appEntityId).orElse(null)
@@ -59,13 +54,11 @@ class AppService(
     return appRepository.findAllByOrganizationIdOrderByNameAsc(organizationId)
   }
 
-  /** How many organizations currently have the app installed. */
   @Transactional(readOnly = true)
   fun countInstalls(appEntityId: Long): Long {
     return appInstallRepository.countByRegisteredAppId(appEntityId)
   }
 
-  /** How many organizations hold each app, in one query, for a list of apps. */
   @Transactional(readOnly = true)
   fun countInstallsByApp(appEntityIds: Collection<Long>): Map<Long, Long> {
     if (appEntityIds.isEmpty()) return emptyMap()
@@ -74,7 +67,6 @@ class AppService(
       .associate { (it[0] as Long) to (it[1] as Long) }
   }
 
-  /** Resolves an app by its app-level `client_id`. The caller must still verify the secret. */
   @Transactional(readOnly = true)
   fun resolveByClientId(clientId: String): App? {
     return appRepository.findByClientId(clientId)
@@ -84,7 +76,6 @@ class AppService(
     return AppSummary(id = app.id, appId = app.appId, name = app.name)
   }
 
-  /** The app-level credentials, disclosed only in the response to registering the app. */
   data class AppCredentials(
     val clientId: String,
     val clientSecret: String,
@@ -92,10 +83,6 @@ class AppService(
   )
 
   companion object {
-    /**
-     * Every successful manifest read is a health check: it stamps the check time and clears any
-     * failure state, whether it came from registration, a refresh, or the periodic sweep.
-     */
     fun markManifestHealthy(
       app: App,
       checkedAt: java.util.Date,
@@ -109,7 +96,6 @@ class AppService(
       app.unhealthyNotifiedAt = null
     }
 
-    /** The wire form of [io.tolgee.model.apps.App.manifestScopes]. */
     fun joinScopes(scopes: Set<io.tolgee.model.enums.Scope>): String =
       scopes.map { it.value }.sorted().joinToString(",")
