+++ backend/security/src/test/kotlin/io/tolgee/security/authorization/ProjectAuthorizationInterceptorTest.kt
@@ -27,6 +27,7 @@ import io.tolgee.fixtures.andIsOk
 import io.tolgee.model.Project
 import io.tolgee.model.UserAccount
 import io.tolgee.model.enums.Scope
+import io.tolgee.security.AppProjectContextBinder
 import io.tolgee.security.OrganizationHolder
 import io.tolgee.security.ProjectContextService
 import io.tolgee.security.ProjectHolder
@@ -36,7 +37,6 @@ import io.tolgee.security.authentication.AuthenticationFacade
 import io.tolgee.security.authentication.ReadOnlyOperation
 import io.tolgee.security.authentication.TolgeeAuthentication
 import io.tolgee.security.authentication.WriteOperation
-import io.tolgee.service.apps.AppEnablementService
 import io.tolgee.service.organization.OrganizationService
 import io.tolgee.service.project.ProjectService
 import io.tolgee.service.security.PermissionService
@@ -93,9 +93,7 @@ class ProjectAuthorizationInterceptorTest {
       projectHolder,
       organizationHolder,
       activityHolder,
-      Mockito.mock(AppEnablementService::class.java),
-      Mockito.mock(PermissionService::class.java),
-      Mockito.mock(UserAccountService::class.java),
+      Mockito.mock(AppProjectContextBinder::class.java),
     )
 
   private val projectAuthenticationInterceptor =
